safari-600.7-branch/LayoutTests/fast/css/ex-unit-with-no-x-height-expected.txt

This test ensures that the CSS 'ex' unit is treated as half the size of the CSS 'em' unit when a font has no x-height information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS measureExBox(); is '50px'
PASS successfullyParsed is true

TEST COMPLETE
